requirements.txt
# Core dependencies
pandas>=2.2.3
numpy>=1.26.4
psycopg2-binary>=2.9.9
pybaseball>=2.2.7
requests>=2.28.0
sqlalchemy>=2.0.0
boto3==1.34.162

# For local testing
python-dotenv>=1.0.0

# AWS Lambda runtime interface (already included in base image, but explicit)
# awslambdaric